+++ go.mod
@@ -9,6 +9,7 @@ require (
 	github.com/google/go-cmp v0.5.5
 	github.com/hashicorp/go-getter v1.5.6
 	github.com/hashicorp/vault/api v1.5.0 // indirect
+	github.com/hashicorp/vault/api/auth/approle v0.1.1 // indirect
 	github.com/jinzhu/now v1.1.1 // indirect
 	github.com/julienschmidt/httprouter v1.3.0
 	github.com/kr/pretty v0.3.0 // indirect
